+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- The release shell's makeRelease now uses the provided github token correctly.
+
 ## [2.0.0] - 2021-12-16
 ### Packages
 - avery: 2.0.0